refactor: Hoist `lock-unlock` to root `src` directory
Align with Gutenberg project practices.
